refactor: disk space management logic
