+++ src/main/java/com/aws/greengrass/logmanager/LogManagerService.java
@@ -27,6 +27,7 @@ import com.aws.greengrass.logmanager.model.LogFile;
 import com.aws.greengrass.logmanager.model.LogFileGroup;
 import com.aws.greengrass.logmanager.model.LogFileInformation;
 import com.aws.greengrass.logmanager.model.ProcessingFiles;
+import com.aws.greengrass.logmanager.services.DiskSpaceManagementService;
 import com.aws.greengrass.util.Coerce;
 import com.aws.greengrass.util.NucleusPaths;
 import com.aws.greengrass.util.Utils;
@@ -39,21 +40,12 @@ import org.slf4j.event.Level;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.nio.file.FileSystems;
-import java.nio.file.Files;
 import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardWatchEventKinds;
-import java.nio.file.WatchEvent;
-import java.nio.file.WatchKey;
-import java.nio.file.WatchService;
-import java.nio.file.Watchable;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -62,14 +54,12 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import javax.inject.Inject;
 
 import static com.aws.greengrass.componentmanager.KernelConfigResolver.CONFIGURATION_CONFIG_KEY;
@@ -116,7 +106,6 @@ public class LogManagerService extends PluginService {
     public static final int DEFAULT_PERIODIC_UPDATE_INTERVAL_SEC = 300;
     public static final int MAX_CACHE_INACTIVE_TIME_SECONDS = 60 * 60 * 24; // 1 day
 
-    private final Object spaceManagementLock = new Object();
     private final List<Consumer<EventType>> serviceStatusListeners = new ArrayList<>();
 
     // public only for integ tests
@@ -125,6 +114,7 @@ public class LogManagerService extends PluginService {
 
     public final Map<String, ProcessingFiles> processingFilesInformation =
             new ConcurrentHashMap<>();
+    private final DiskSpaceManagementService diskSpaceManagementService;
 
 
     @Getter
@@ -136,7 +126,6 @@ public class LogManagerService extends PluginService {
     private final AtomicBoolean isCurrentlyUploading = new AtomicBoolean(false);
     @Getter
     private int periodicUpdateIntervalSec;
-    private Future<?> spaceManagementThread;
     private final Path workDir;
 
     /**
@@ -154,6 +143,7 @@ public class LogManagerService extends PluginService {
         this.uploader = uploader;
         this.logsProcessor = logProcessor;
         this.executorService = executorService;
+        this.diskSpaceManagementService = new DiskSpaceManagementService(this.executorService);
         this.workDir = nucleusPaths.workPath(LOGS_UPLOADER_SERVICE_TOPICS);
 
         topics.lookupTopics(CONFIGURATION_CONFIG_KEY).subscribe((why, newv) -> {
@@ -253,7 +243,7 @@ public class LogManagerService extends PluginService {
         this.componentLogConfigurations = newComponentLogConfigurations;
         logger.atInfo().log("Finished processing LogManager configuration");
 
-        scheduleSpaceManagementThread();
+        this.diskSpaceManagementService.start(this.componentLogConfigurations);
     }
 
     private void handleUserComponentConfiguration(Object componentConfigObject,
@@ -357,29 +347,6 @@ public class LogManagerService extends PluginService {
             }
         });
         setDiskSpaceLimit(diskSpaceLimitString.get(), diskSpaceLimitUnitString.get(), componentLogConfiguration);
-    }
-
-    @SuppressWarnings("PMD.AvoidCatchingThrowable")
-    private void scheduleSpaceManagementThread() {
-        synchronized (spaceManagementLock) {
-            if (spaceManagementThread != null) {
-                spaceManagementThread.cancel(true);
-            }
-            spaceManagementThread = this.executorService.submit(() -> {
-                try {
-                    logger.atInfo().log("Starting space management thread");
-                    startWatchServiceOnLogFilePaths();
-                } catch (IOException e) {
-                    //TODO: fail the deployment?
-                    logger.atError().cause(e).log("Unable to start space management thread.");
-                } catch (InterruptedException e) {
-                    Thread.currentThread().interrupt();
-                } catch (Throwable e) {
-                    logger.atError().log("Failure in log manager space management", e);
-                    scheduleSpaceManagementThread(); // restart space management
-                }
-            });
-        }
     }
 
     private void setMinimumLogLevel(String minimumLogLevel, ComponentLogConfiguration componentLogConfiguration) {
@@ -764,145 +731,6 @@ public class LogManagerService extends PluginService {
     public void startup() throws InterruptedException {
         super.startup();
         processLogsAndUpload();
-    }
-
-    /**
-     * Starts a Watch Service on all the log directories where customer has specified disk space limit.
-     * If any file is modified and created in any of those directories, we get an event containing that information.
-     * The log manager will then verify that the size of all log files in that directory for the component does not
-     * exceed the disk space limit set by the customer. If it does, the log manager will delete the log files until
-     * the limit is met.
-     *
-     * @throws IOException          if unable to initialise a new Watch Service.
-     * @throws InterruptedException if thread is shutdown
-     */
-    private void startWatchServiceOnLogFilePaths() throws IOException, InterruptedException {
-        //TODO: Optimize this.
-        // The optimization would be to have best of both worlds. The file watcher will mark the changed components
-        // log directories. Another scheduled thread will look at that and clean up files if necessary.
-        try (WatchService watchService = FileSystems.getDefault().newWatchService()) {
-            try {
-                componentLogConfigurations.forEach((componentName, componentLogConfiguration) -> {
-                    // Only register the path of a component if the disk space limit is set.
-                    if (componentLogConfiguration != null && componentLogConfiguration.getDiskSpaceLimit() != null
-                            && componentLogConfiguration.getDiskSpaceLimit() > 0) {
-                        Path path = componentLogConfiguration.getDirectoryPath();
-                        try {
-                            path.register(watchService, StandardWatchEventKinds.ENTRY_CREATE,
-                                    StandardWatchEventKinds.ENTRY_MODIFY);
-                            logger.atDebug().kv("filePath", path).log("Start watching file for log space management.");
-                        } catch (IOException e) {
-                            logger.atError().cause(e).log("Unable to watch {} for log space management.",
-                                    path);
-                        }
-                    }
-                });
-
-                while (true) {
-                    final WatchKey watchKey = watchService.take();
-                    final Watchable watchable = watchKey.watchable();
-
-                    //Since we are registering only paths in thw watch service, the watchables must be paths
-                    if (watchable instanceof Path) {
-                        final Path directory = (Path) watchable;
-                        Map<String, ComponentLogConfiguration> updatedComponentsConfiguration = new HashMap<>();
-                        // Based on the directory and file name created/modified, the log manager will store the
-                        // component information which need to be checked.
-                        List<ComponentLogConfiguration> allComponentsConfiguration =
-                                componentLogConfigurations.values().stream()
-                                        .filter(componentLogConfiguration ->
-                                                componentLogConfiguration.getDirectoryPath().equals(directory))
-                                        .collect(Collectors.toList());
-
-                        // The events can have multiple log files in the same directory. We need to make sure we get the
-                        // correct component based on the file name pattern.
-                        for (WatchEvent<?> event : watchKey.pollEvents()) {
-                            String fileName = Coerce.toString(event.context());
-                            if (fileName == null) {
-                                continue;
-                            }
-                            List<ComponentLogConfiguration> list = allComponentsConfiguration.stream()
-                                    .filter(componentLogConfiguration -> componentLogConfiguration
-                                            .getFileNameRegex().matcher(fileName).find())
-                                    .collect(Collectors.toList());
-                            list.forEach(componentLogConfiguration ->
-                                    updatedComponentsConfiguration.putIfAbsent(componentLogConfiguration.getName(),
-                                            componentLogConfiguration));
-                        }
-
-                        // Get the total log files size in the directories for all updates components. If the log files
-                        // size exceed the limit set by the customer, the log manager will figure out the minimum bytes
-                        // to be deleted inorder to meet the limit. It will then go ahead and delete the oldest files
-                        // until the minimum bytes have been deleted.
-                        for (ComponentLogConfiguration componentLogConfiguration :
-                                updatedComponentsConfiguration.values()) {
-                            try {
-                                deleteFilesIfNecessary(componentLogConfiguration);
-                            } catch (UncheckedIOException e) {
-                                logger.atWarn().log("Unchecked error thrown when collecting files to be deleted",
-                                        Utils.getUltimateCause(e));
-                            }
-                        }
-                    }
-
-                    if (!watchKey.reset()) {
-                        logger.atError().log("Log Space Management encountered an issue. Returning.");
-                        scheduleSpaceManagementThread();
-                        break;
-                    }
-                }
-            } catch (IOException e) {
-                // If there is any other IOException, then we should restart the thread.
-                scheduleSpaceManagementThread();
-            }
-        }
-    }
-
-    /**
-     * Deletes the oldest log files which matches the file name pattern. The log files will be deleted until the
-     * minimum number of bytes have been deleted.
-     *
-     * @param componentLogConfiguration Log configuration to investigate for deletion
-     * @throws IOException for errors during directory walking
-     */
-    private void deleteFilesIfNecessary(ComponentLogConfiguration componentLogConfiguration) throws IOException {
-        if (componentLogConfiguration.getDiskSpaceLimit() == null) {
-            return;
-        }
-        try (Stream<Path> fileStream = Files.walk(componentLogConfiguration.getDirectoryPath())
-                .filter(p -> {
-                    File file = p.toFile();
-                    return file.isFile() && componentLogConfiguration.getFileNameRegex()
-                            .matcher(file.getName()).find();
-                })) {
-            List<Path> paths = fileStream.collect(Collectors.toList());
-            long totalBytes = paths.stream().mapToLong(p -> p.toFile().length()).sum();
-            long minimumBytesToBeDeleted = totalBytes - componentLogConfiguration.getDiskSpaceLimit();
-
-            // If we don't need to remove any bytes, or if the file count is only 1 (or less), then there's nothing
-            // to do.
-            if (minimumBytesToBeDeleted <= 0 || paths.size() < 2) {
-                return;
-            }
-
-            long bytesDeleted = 0;
-            // Sort the files by the last modified time.
-            paths.sort(Comparator.comparingLong((p) -> p.toFile().lastModified()));
-            int fileIndex = 0;
-            // stop before the end to skip the active file which should have the newest modified time
-            while (bytesDeleted < minimumBytesToBeDeleted && fileIndex < paths.size() - 1) {
-                Path fileToBeDeleted = paths.get(fileIndex++);
-                long fileSize = fileToBeDeleted.toFile().length();
-                try {
-                    Files.deleteIfExists(fileToBeDeleted);
-                } catch (IOException e) {
-                    logger.atWarn().log("Unable to delete file: {}", fileToBeDeleted.toAbsolutePath(), e);
-                    break;
-                }
-                logger.atInfo().log("Successfully deleted file: {}", fileToBeDeleted.toAbsolutePath());
-                bytesDeleted += fileSize;
-            }
-        }
     }
 
     @Override
